public/python/requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
fuzzywuzzy==0.18.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.15.0
google-cloud-videointelligence==2.11.3
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
mysql-connector-python==8.3.0
numpy==1.24.4
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
urllib3==2.2.1
